Handle LTS releases
Until now, LTS releases were handled as maintenance releases as they
would become LTS when the extended maitenance would have started.

This is no longer the case and we need to take extra care of it,
especially to push the documentation in a versioned directory and the
quickstarts to a versioned branch.